whisper-ai-service/requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
langdetect
